% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbSummary.r
\name{dbSummary}
\alias{dbSummary}
\title{Database summary function}
\usage{
dbSummary()
}
\arguments{
\item{No}{parameters for this function}
}
\description{
This function creates a summary of the contents of the Avian Diet Database
}
\examples{
dbSummary()
}
\keyword{summary}
