% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietSummary.r
\name{dietSummary}
\alias{dietSummary}
\title{Flexible diet summary function}
\usage{
dietSummary(
  commonName,
  by = "Order",
  dietType = "Items",
  season = NULL,
  region = NULL,
  yearRange = c(1700, 2100),
  db = NULL
)
}
\arguments{
\item{commonName}{common name of bird species to summarize; case-insensitive.}

\item{by}{taxonomic level of prey to summarize diet by; possibile values include 'Kingdom', 'Phylum', 'Class', 'Order', 'Family', 'Genus', or 'Species'; defaults to 'Order'.}

\item{dietType}{the type of quantitative diet data; possible values include percent by 'Items', 'Wt_or_Vol' (weight or volume), or percent 'Occurrence'; defaults to 'Items'.}

\item{season}{the season for which a diet summary should be conducted; possible values include 'spring', 'summer', 'fall', 'winter', or 'any'; defaults to 'any'.}

\item{region}{the region for which a diet summary should be conducted; typically these are US or Mexican state or Canadian province names.}

\item{yearRange}{a vector specifying the minimum and maximum years over which a diet summary should be conducted; by default all years will be included.}

\item{db}{used to specify an alternative Diet Database object, mainly for testing; the default results in the main 'dietdb' data object being used.}
}
\description{
This function provides a summary of quantitative diet data for a given bird species, taxonomic level of prey resolution, data type, season, region, and year range
}
\examples{
dietSummary("Bald Eagle")
dietSummary("Bald Eagle", by = "Class", season = "winter", dietType = "Items", yearRange = c(1985, 2000))
}
\keyword{diet}
\keyword{summary}
