% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietSummaryByPrey.r
\name{dietSummaryByPrey}
\alias{dietSummaryByPrey}
\title{Summarize the importance of a single prey item across bird species}
\usage{
dietSummaryByPrey(
  preyName,
  preyLevel,
  larvaOnly = FALSE,
  dietType = NULL,
  season = NULL,
  region = NULL,
  yearRange = c(1700, 2100),
  speciesMean = FALSE
)
}
\arguments{
\item{preyName}{scientific name of prey item to summarize, which may be at any taxonomic level denoted by preyLevel}

\item{preyLevel}{taxonomic level of prey; possibile values include 'Kingdom', 'Phylum', 'Class', 'Order', 'Family', 'Genus', or 'Species'}

\item{larvaOnly}{if TRUE, only records in which the specified prey taxon was consumed as a larva are returned; if FALSE (default), records for any life stage are returned.}

\item{dietType}{the way in which diet data were quantified; possible values include percent by 'Items', 'Wt_or_Vol' (weight or volume), or 'Occurrence'; defaults to 'Items'.}

\item{season}{the season for which a diet summary should be conducted; possible values include 'spring', 'summer', 'fall', 'winter', or 'any'; defaults to 'any'.}

\item{region}{the region for which a diet summary should be conducted; typically these are US or Mexican state or Canadian province names; by default all regions will be included.}

\item{yearRange}{a vector specifying the minimum and maximum years over which a diet summary should be conducted; by default all years will be included.}

\item{speciesMean}{logical value indicating whether to average across all regions, years and seasons to yield a single diet value per bird species-Diet_Type combination; default is FALSE in which case all diet analyses in which the prey taxon appears will be listed separately.}
}
\description{
This function provides a summary of quantitative diet data for a given prey type
}
\examples{
dietSummaryByPrey("Lepidoptera", preyLevel = "Order", larvaOnly = TRUE, dietType = "Items", season = "spring")
}
\keyword{diet}
\keyword{prey}
\keyword{summary}
